+++ src/common/enums/role.enum.ts
@@ -1,0 +1,6 @@
+export enum RoleType {
+  SUPERADMIN = 'superadmin',
+  ADMIN = 'admin',
+  HOST = 'host',
+  USER = 'user',
+}