+++ src/modules/users/users.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { User } from './entities/user.entity';
+import { UsersService } from './users.service';
+import { Role } from './entities/role.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([User, Role])],
+  providers: [UsersService],
+  exports: [UsersService],
+})
+export class UsersModule {}